refactor: add useEtherscanOpener to replace useEtherscanOpen hook
