--- modules/blockChain/hooks/useIsMultisig.ts
@@ -1,10 +1,0 @@
-import { useWeb3 } from './useWeb3'
-import { useIsContract } from './useIsContract'
-
-export const useIsMultisig = () => {
-  const { walletAddress } = useWeb3()
-  const { data: isContract, isLoading: loading } = useIsContract(
-    walletAddress ?? undefined,
-  )
-  return [isContract, loading]
-}
