+++ modules/blockChain/hooks/useSendTransactionGnosisWorkaround.ts
@@ -2,12 +2,11 @@ import { useCallback } from 'react'
 import { useWeb3 } from 'modules/blockChain/hooks/useWeb3'
 import { PopulatedTransaction } from '@ethersproject/contracts'
 import { sendTransactionGnosisWorkaround } from '../utils/sendTransactionGnosisWorkaround'
-import { useIsMultisig } from './useIsMultisig'
+import { useIsContract } from './useIsContract'
 
 export function useSendTransactionGnosisWorkaround() {
-  const { web3Provider } = useWeb3()
-  // TODO: track loading state of this swr in the ui on yes/no/enact button
-  const [isMultisig] = useIsMultisig()
+  const { walletAddress, web3Provider } = useWeb3()
+  const { data: isMultisig } = useIsContract(walletAddress)
 
   return useCallback(
     (tx: PopulatedTransaction) =>
