+++ modules/blockChain/hooks/useTransactionSender.ts
@@ -6,7 +6,7 @@ import { ResultTx, TxStatus } from '../types'
 import { PopulatedTransaction } from '@ethersproject/contracts'
 import { openWindow } from 'modules/shared/utils/openWindow'
 import { getGnosisSafeLink } from '../utils/getGnosisSafeLink'
-import { getEtherscanLink } from '../utils/getEtherscanLink'
+import { getEtherscanLink } from '../utils/etherscan'
 import { getErrorMessage } from 'modules/shared/utils/getErrorMessage'
 import { ToastError } from '@lidofinance/lido-ui'
