+++ modules/blockChain/utils/etherscan.ts
@@ -1,11 +1,7 @@
 import invariant from 'tiny-invariant'
 import { CHAINS } from '../chains'
 
-export enum ETHERSCAN_ENTITIES {
-  TX = 'tx',
-  TOKEN = 'token',
-  ADDRESS = 'address',
-}
+export type EtherscanEntity = 'tx' | 'token' | 'address'
 
 const ETHERSCAN_PREFIX_BY_NETWORK: Partial<Record<CHAINS, string>> = {
   [CHAINS.Mainnet]: '',
@@ -22,28 +18,27 @@ export const getEtherscanPrefix = (chainId: CHAINS) => {
 export const getEtherscanLink = (
   chainId: CHAINS,
   hash: string,
-  entity: string,
+  entity: EtherscanEntity,
 ) => {
   const prefix = getEtherscanPrefix(chainId)
   invariant(hash && typeof hash === 'string', 'Hash should be a string')
-  invariant(entity && typeof entity === 'string', 'Entity should be a string')
   return `https://${prefix}etherscan.io/${entity}/${hash}`
 }
 
 export const getEtherscanTxLink = (chainId: number, hash: string): string => {
-  return getEtherscanLink(chainId, hash, ETHERSCAN_ENTITIES.TX)
+  return getEtherscanLink(chainId, hash, 'tx')
 }
 
 export const getEtherscanTokenLink = (
   chainId: number,
   hash: string,
 ): string => {
-  return getEtherscanLink(chainId, hash, ETHERSCAN_ENTITIES.TOKEN)
+  return getEtherscanLink(chainId, hash, 'token')
 }
 
 export const getEtherscanAddressLink = (
   chainId: number,
   hash: string,
 ): string => {
-  return getEtherscanLink(chainId, hash, ETHERSCAN_ENTITIES.ADDRESS)
+  return getEtherscanLink(chainId, hash, 'address')
 }
