+++ modules/delegation/ui/DelegationAddressPop/DelegationAddressPop.tsx
@@ -18,7 +18,6 @@ import {
 import { calcPopoverPosition } from './calcPopoverPosition'
 import { getPublicDelegateByAddress } from 'modules/delegation/utils/getPublicDelegateName'
 import { PublicDelegateAvatar } from '../PublicDelegateAvatar'
-import { ETHERSCAN_ENTITIES } from 'modules/blockChain/utils/getEtherscanLink'
 
 type IdenticonBadgeProps = React.ComponentProps<typeof IdenticonBadge>
 
@@ -75,10 +74,7 @@ export function DelegationAddressPop({ children, ...badgeProps }: Props) {
                 onClick={handleOpen}
               />
             </BadgeWrap>
-            <CopyOpenActions
-              value={address}
-              entity={ETHERSCAN_ENTITIES.ADDRESS}
-            />
+            <CopyOpenActions value={address} entity="address" />
           </div>
           {delegateAddress && (
             <div>
@@ -107,10 +103,7 @@ export function DelegationAddressPop({ children, ...badgeProps }: Props) {
                   />
                 )}
               </BadgeWrap>
-              <CopyOpenActions
-                value={delegateAddress}
-                entity={ETHERSCAN_ENTITIES.ADDRESS}
-              />
+              <CopyOpenActions value={delegateAddress} entity="address" />
             </div>
           )}
         </Pop>
