+++ modules/shared/ui/Common/AddressPop/AddressPop.tsx
@@ -7,7 +7,6 @@ import { IdenticonBadge } from '@lidofinance/lido-ui'
 import { Wrap, Pop, BadgeWrap } from './AddressPopStyle'
 
 import { calcPopoverPosition } from './calcPopoverPosition'
-import { ETHERSCAN_ENTITIES } from 'modules/blockChain/utils/getEtherscanLink'
 
 type Props = React.ComponentProps<typeof IdenticonBadge>
 
@@ -60,10 +59,7 @@ export function AddressPop({ children, ...badgeProps }: Props) {
               onClick={handleOpen}
             />
           </BadgeWrap>
-          <CopyOpenActions
-            value={address}
-            entity={ETHERSCAN_ENTITIES.ADDRESS}
-          />
+          <CopyOpenActions value={address} entity="address" />
         </Pop>
       )}
     </Wrap>
